# doc/generated/openapi.yaml
---
openapi: 3.1.0
components:
  schemas:
    Adr:
      type: object
      required:
      - name
      - content
      properties:
        name:
          type: string
        content:
          type: string
    Api:
      type: object
      required:
      - name
      - type
      - lifecycle
      - description
      - definitionUrl
      properties:
        name:
          type: string
        type:
          type:
          - string
          - "null"
        lifecycle:
          type:
          - string
          - "null"
        description:
          type:
          - string
          - "null"
        definitionUrl:
          type:
          - string
          - "null"
    Component:
      type: object
      required:
      - name
      - owner
      - type
      - lifecycle
      - description
      - tags
      - links
      - annotations
      - providesApis
      properties:
        name:
          type: string
        owner:
          type:
          - string
          - "null"
        type:
          type:
          - string
          - "null"
        lifecycle:
          type:
          - string
          - "null"
        description:
          type:
          - string
          - "null"
        tags:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/Link"
        annotations:
          type: object
          additionalProperties:
            type: string
        providesApis:
          type: array
          items:
            type: string
    Graph:
      type: object
      required:
      - nodes
      - edges
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
    GraphEdge:
      type: object
      required:
      - id
      - source
      - target
      - relation
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        relation:
          type: string
    GraphNode:
      type: object
      required:
      - id
      - label
      - type
      - kind
      - owner
      - lifecycle
      - isFocus
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
        kind:
          type: string
        owner:
          type:
          - string
          - "null"
        lifecycle:
          type:
          - string
          - "null"
        isFocus:
          type: boolean
    Link:
      type: object
      required:
      - url
      - title
      properties:
        url:
          type: string
        title:
          type:
          - string
          - "null"
    TechDoc:
      type: object
      required:
      - name
      - content
      properties:
        name:
          type: string
        content:
          type: string
paths:
  /api/catalog/apis:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Api"
      summary: Get All Apis
      tags:
      - Catalog Controller
  /api/catalog/components:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Component"
      summary: Get Components
      tags:
      - Catalog Controller
  /api/catalog/components/{name}:
    get:
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Component"
      summary: Get Component
      tags:
      - Catalog Controller
  /api/catalog/components/{name}/adrs:
    get:
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Adr"
      summary: Get Adrs
      tags:
      - Catalog Controller
  /api/catalog/components/{name}/apis:
    get:
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Api"
      summary: Get Apis
      tags:
      - Catalog Controller
  /api/catalog/components/{name}/docs:
    get:
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TechDoc"
      summary: Get Docs
      tags:
      - Catalog Controller
  /api/catalog/components/{name}/docs/assets/{assetPath}:
    get:
      parameters:
      - name: assetPath
        in: path
        required: true
        schema:
          type: string
          pattern: .+
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*': {}
      summary: Get Doc Asset
      tags:
      - Catalog Controller
  /api/catalog/components/{name}/graph:
    get:
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Graph"
      summary: Get Graph
      tags:
      - Catalog Controller
  /api/techradar:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Get Tech Radar
      tags:
      - Tech Radar Controller
info:
  title: centerstage API
  version: unspecified
servers:
- url: http://localhost:50700
  description: Auto generated value
- url: http://0.0.0.0:50700
  description: Auto generated value